Refactor DFA page structure
